fixing validation of SFU CSSS Google Drive "CSSS" folder. Combines and cleans up code from PRs:  * https://github.com/CSSS/csss-site/pull/221  * https://github.com/CSSS/csss-site/pull/222  * https://github.com/CSSS/csss-site/pull/223  * https://github.com/CSSS/csss-site/pull/224  * https://github.com/CSSS/csss-site/pull/225  * https://github.com/CSSS/csss-site/pull/226
Changes
 * more logging added so that errors are easier to trace to their cause
in the logging
 * separates the logic for fixing the permissions for the root folder
and the logic for fixing the ownership and permissions of all the
subfiles/folders
 * restructures code to email the owners of folders with the folders
whose ownership needs to be changed. This is done because the code was
previously attempting to make a comment on the folders as it assumed
that anything that can have its ownership changed and also be commented
on.
 * the code will no longer attempt to delete an uploaded file owned by
someone else as it seems like the API doesn't really have that functionality.
It will now instead just duplicate the file, name the original file to
indicate it should no longer be used and add a comment to the original
to indicate to its owner that it needs to be deleted
 * Adds Gmail class so that the gdrive_api can send emails to the owners
of the folders that need to have their ownership updated. This change
also necessitated that the functionality of the code that sends the
introductory email to the new officers also use this new object

Author:    Jace Manshadi <jmanshad@protonmail.com>